+++ compiler/src/dotty/tools/dotc/decompiler/DecompilationPrinter.scala
@@ -8,7 +8,7 @@ import scala.io.Codec
 import dotty.tools.dotc.core.Contexts._
 import dotty.tools.dotc.core.Phases.Phase
 import dotty.tools.dotc.core.tasty.TastyPrinter
-import dotty.tools.dotc.tastyreflect.ReflectionImpl
+import dotty.tools.dotc.quoted.reflect.ReflectionImpl
 import dotty.tools.io.File
 
 /** Phase that prints the trees in all loaded compilation units.
