+++ compiler/src/dotty/tools/dotc/quoted/MacroExpansion.scala
@@ -1,4 +1,4 @@
-package dotty.tools.dotc.tastyreflect
+package dotty.tools.dotc.quoted
 
 import dotty.tools.dotc.ast.tpd
 import dotty.tools.dotc.core._
