+++ compiler/src/dotty/tools/dotc/transform/ReifyQuotes.scala
@@ -21,7 +21,7 @@ import scala.collection.mutable
 import dotty.tools.dotc.core.Annotations._
 import dotty.tools.dotc.core.Names._
 import dotty.tools.dotc.core.StdNames._
-import dotty.tools.dotc.core.quoted._
+import dotty.tools.dotc.quoted._
 import dotty.tools.dotc.transform.TreeMapWithStages._
 import dotty.tools.dotc.typer.Inliner
