+++ compiler/src/dotty/tools/dotc/transform/TreeMapWithStages.scala
@@ -7,7 +7,6 @@ import dotty.tools.dotc.config.Printers.staging
 import dotty.tools.dotc.core.Constants._
 import dotty.tools.dotc.core.Decorators._
 import dotty.tools.dotc.core.Flags._
-import dotty.tools.dotc.core.quoted._
 import dotty.tools.dotc.core.NameKinds._
 import dotty.tools.dotc.core.Types._
 import dotty.tools.dotc.core.Contexts._
@@ -15,6 +14,7 @@ import dotty.tools.dotc.core.StagingContext._
 import dotty.tools.dotc.core.StdNames._
 import dotty.tools.dotc.core.Symbols._
 import dotty.tools.dotc.core.tasty.TreePickler.Hole
+import dotty.tools.dotc.quoted._
 import dotty.tools.dotc.util.Spans._
 import dotty.tools.dotc.util.Property
 import dotty.tools.dotc.transform.SymUtils._
